BTP project
The problem that we have chosen to solve is navigation path planning from a source to a set of destinations. This problem in other words is a Travelling Salesperson problem. Now, solving this problem has many advantages in many areas like planning, logistics, and manufacture of microchips. It is a sub-problem in many areas such as DNA sequencing…